# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AutoMarket API",
        "description": "API for AutoMarket platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Base server (prefix /api in paths)"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User registered"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User logged in"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "User logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/settings/public": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "operationId": "648312a4b4d172b162ca8689ddf5f949",
                "responses": {
                    "200": {
                        "description": "Public settings and exchange rates"
                    }
                }
            }
        },
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "727634867f59796290a2fff276979198",
                "parameters": [
                    {
                        "name": "make",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year_min",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year_max",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mileage_min",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mileage_max",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fuel_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transmission",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicles list"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "operationId": "a88e0bca4193c91e713d0441659d199c",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "operationId": "0185374b4b707b85ca8a97a6f5f366e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "operationId": "c4d991b5c398a7ccd4e73f9a81c355d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "operationId": "3442f99c22f47affb335443ef012e712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vehicle deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication endpoints"
        },
        {
            "name": "Settings",
            "description": "Settings endpoints"
        },
        {
            "name": "Vehicles",
            "description": "Vehicle listing and management"
        }
    ]
}